requirements.txt
charset-normalizer==2.0.12
cmake
dlib
face-alignment==1.3.5
h5py
idna==3.3
imageio==2.18.0
imutils
kornia==0.6.4
llvmlite==0.38.0
networkx==2.8
numba==0.55.1
numpy
opencv-python==4.3.0.36
packaging==21.3
pandas
pillow==9.1.1
piq
py
pyparsing==3.0.8
pyqt5==5.15.6
pyqt5-qt5==5.15.2
pyqt5-sip==12.10.1
pywavelets==1.3.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
sqlalchemy
tifffile==2022.4.28
torch
torchvision
typing-extensions==4.2.0
urllib3==1.26.9
wandb
